requirements.txt
pandas
python-dotenv
streamlit
langchain-core
langchain-community
langchain-openai
openai
tqdm
